requirements.txt
Pillow==9.3.0
# torch==2.3.0+cu118
torch==2.3.0
torchaudio==2.3.0
torchdata==0.7.1
torchtext==0.18.0
torchvision==0.18.0
tqdm==4.66.1
opencv-python==4.10.0.84
matplotlib==3.7.2